requirements.txt
fpdf==1.7.2
numpy==1.15.4
opencv-contrib-python==3.4.3.18
Pillow==5.3.0